wq executor: send and set X509 user proxy
adds x509_proxy keyword to wq executor arguments to indicate the local
path of the X509 proxy to be sent along the tasks.

If set to None (the default), use the value of the X509_USER_PROXY
environment variable if set, or the file /tmp/x509up_u${UID} if exits.
If False, no proxy is sent.