wq: adds log of accumulated chunks
As chunks get accumulated, they are written to a log amenable to pandas
processing. The log has information about the resources used, the
dataset involved, and the source file.

The current motivation is to debug processors in topcoffea for which
certain datasets/filenames seem to be taking much longer than others.

current columns are:

id,category,status,dataset,file,range_start,range_stop,accum_parent,\
time_start,time_end,cpu_time,memory

id is the work queue task id.

category is one of preprocessing, processing, or accumulating.

accum_parent is the work queue task id of the task that accumulated the
results. The final accumulating task has an accum_parent of 0.

For preprocessing tasks, range_start is always 0, and range_stop is the
number of events in the file.

For accumulating tasks, range_start is always 0, and range_stop is the
number of events accumulated. dataset and file are empty strings.