wq: option to activate ssl between manager and workers
The WorkQueueExecutor accepts the parameter `ssl`. If `ssl=True`, then
temporary autosign key and certs are generated and used. Otherwise,
ssl=(key,cert) should indicate the paths to a key and cert in pem
format.

When activating ssl in the wq manager, workers should be launched with
the --ssl command line argument.