Added support for multiple systematic variations of a single weight
This covers the need for multiple systematic variation of a single weight added to the event weight.
E.g. the various btag SF variations in CMS.

A list of modifierNames is added as argument to the function `add_multivariation`.
The final modifier to the event name is formatted as `name_modifierUp/Down`.

The corresponding pytest has been included.